Update
Added connect/disconnect functionality on the create and update category.
Added a separate API for updating category show state.
Added reset to default button when updating a collection or category
Fixed isAdmin on getCategories
Added padding bottom for item details page to have the ability to press add to cart when using on mobile
More to come later...