update
When a product is deleted, its images are deleted from the database(firebase & hygraph).
Also, its tags, variants, and order-items.
Even orders that got emptied of order-items are deleted. (needs more testing to check if all orders with no items are deleted or only recently emptied ones.)
Fixed styling of submitting on the cart page.
If image is not found in the cart, it won't crash.
Added scroll buttons in products and orders pages for admin.